+++ DioriteAPI/src/main/java/org/diorite/material/blocks/earth/FarmlandMat.java
@@ -20,7 +20,7 @@ public class FarmlandMat extends EarthMat
     /**
      * Sub-ids used by diorite/minecraft by default
      */
-    public static final byte USED_DATA_VALUES = 8;
+    public static final int USED_DATA_VALUES = 8;
 
     public static final FarmlandMat FARMLAND_UNHYDRATED = new FarmlandMat();
     public static final FarmlandMat FARMLAND_MOISTURE_1 = new FarmlandMat("MOISTURE_1", 1);
