+++ DioriteAPI/src/main/java/org/diorite/material/blocks/others/IronTrapdoorMat.java
@@ -17,7 +17,7 @@ public class IronTrapdoorMat extends TrapdoorMat
     /**
      * Sub-ids used by diorite/minecraft by default
      */
-    public static final byte  USED_DATA_VALUES = 16;
+    public static final int USED_DATA_VALUES = 16;
 
     public static final IronTrapdoorMat IRON_TRAPDOOR_WEST_BOTTOM  = new IronTrapdoorMat();
     public static final IronTrapdoorMat IRON_TRAPDOOR_SOUTH_BOTTOM = new IronTrapdoorMat(BlockFace.SOUTH, false, false);
